[refactor] split specs into rag.server.models
