+++ libs/rag/server/chat/kb_chat.py
@@ -4,7 +4,7 @@ from fastapi import Body
 from rag.server.chat.llm.base import LLMFactory
 from rag.server.chat.utils import construct_message
 from rag.server.kb.kb_api import search
-from rag.server.specs import History
+from rag.server.models.chat_spec import History
 from rag.settings import Settings
 from rag.utils import build_logger
