+++ libs/rag/server/models/api_spec.py
@@ -1,0 +1,30 @@
+from typing import Any, List
+
+from rag.server.pydantic_v2 import BaseModel, Field
+
+
+class BaseResponse(BaseModel):
+    code: int = Field(200, description="API status code")
+    msg: str = Field("success", description="API status message")
+    data: Any = Field(None, description="API data")
+
+    class Config:
+        json_schema_extra = {
+            "example": {
+                "code": 200,
+                "msg": "success",
+            }
+        }
+
+
+class ListResponse(BaseResponse):
+    data: List[Any] = Field(..., description="List of data")
+
+    class Config:
+        json_schema_extra = {
+            "example": {
+                "code": 200,
+                "msg": "success",
+                "data": ["doc1.docx", "doc2.pdf", "doc3.txt"],
+            }
+        }
